+++ modules/hanna-react/src/_abstract/_TogglerGroup.tsx
@@ -7,13 +7,16 @@ import { useDomid } from '../utils/useDomid.js';
 
 import { TogglerInputProps } from './_TogglerInput.js';
 
-export type TogglerGroupOption<T = 'default'> = {
+export type TogglerGroupOption<T = 'default', Extras = {}> = {
   value: string;
   label?: T extends 'default' ? string | JSX.Element : T;
   disabled?: boolean;
   id?: string;
-};
-export type TogglerGroupOptions<T = 'default'> = Array<TogglerGroupOption<T>>;
+} & Partial<Extras>;
+
+export type TogglerGroupOptions<T = 'default', Extras = {}> = Array<
+  TogglerGroupOption<T, Extras>
+>;
 
 type RestrictedInputProps = Omit<
   HTMLProps<'input'>,
@@ -28,8 +31,8 @@ type RestrictedInputProps = Omit<
   | 'children'
 >;
 
-export type TogglerGroupProps<T = 'default'> = {
-  options: Array<string> | TogglerGroupOptions<T>;
+export type TogglerGroupProps<T = 'default', Extras = {}> = {
+  options: Array<string> | TogglerGroupOptions<T, Extras>;
   className?: string;
   name?: string;
   disabled?: boolean | ReadonlyArray<number>;
@@ -40,7 +43,7 @@ export type TogglerGroupProps<T = 'default'> = {
     /** The new checked state of the selected value */
     checked: boolean;
     /** The option object being selected */
-    option: TogglerGroupOption<T>;
+    option: TogglerGroupOption<T, Extras>;
     /** The updated value array */
     selectedValues: Array<string>;
   }) => void;