+++ modules/visual-tests/src/routes/test/Multiselect.tsx
@@ -27,13 +27,19 @@ const opts = [
   'Fig',
 ];
 const optsLong = opts.flatMap((val) => [val, `${val} (crate)`]);
-const optsShort = opts.slice(0, 4);
+const testGrouping = true as boolean;
+const optsShort = !testGrouping
+  ? opts.slice(0, 4)
+  : ['Apricot', 'Blueberry']
+      .concat(opts.slice(0, 4))
+      .map((value, i) => ({ value, group: i < 2 ? '' : undefined }));
 const optsFew = opts.slice(5, 7);
 
 const multiselects: Record<string, MultiselectProps> = {
   empty: {
     label: 'Empty',
     options: optsShort,
+    forceSearchable: true,
   },
   small: {
     label: 'Small',